modified the delete route to ensure the correct user can delete sauce

+++ backend/controllers/sauce.js
@@ -53,18 +53,31 @@ exports.updateSauce = (req, res, next) => {
 };
 
 exports.deleteSauce = (req, res, next) => {
-  Sauce.deleteOne({ _id: req.params.id })
-    .then(() => {
-      res.status(200).json({
-        message: "Sauce deleted successfully!",
+  Sauce.findOne({ _id: req.params.id }).then((sauce) => {
+    if (!sauce) {
+      return res.status(404).json({
+        error: new Error("Sauce not found!"),
       });
-    })
-    .catch((error) => {
-      res.status(400).json({
-        error: error,
+    }
+    if (sauce.userId !== req.auth.userId) {
+      return res.status(403).json({
+        error: new Error("Unauthorized request!"),
       });
-    });
+    }
+    Sauce.deleteOne({ _id: req.params.id })
+      .then(() => {
+        res.status(200).json({
+          message: "Sauce deleted successfully!",
+        });
+      })
+      .catch((error) => {
+        res.status(400).json({
+          error: error,
+        });
+      });
+  });
 };
+    
 
 exports.getAllSauces = (req, res, next) => {
     Sauce.find()
